requirements.txt
numpy
cython
scipy
nose
cgen
codepy
coverage
decorator
lockfile
matplotlib
pymongo
pytools
wsgiref
fabric
shovel
joblib
networkx
pillow
skdata
hyperopt
scikit-image
theano
